requirements.txt
openai
requests
langsmith 
langchain
langchain-openai
python-dotenv